# hyperv-agent-openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Hyper-V Agent API",
    "version": "v1",
    "description": "REST API exposing comprehensive Hyper-V host functions (HCS, HCN, WMI, VHD, Storage Management, Jobs).",
    "contact": {
      "name": "HyperV Agent",
      "email": "admin@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "paths": {
...
    "/api/v1/vms/{name}/storage/devices": {
      "get": {
        "summary": "Get VM Storage Devices",
        "description": "Gets list of storage devices in a VM.",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StorageDeviceResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "VM not found"
          }
        }
      },
      "post": {
        "summary": "Add VM Storage Device",
        "description": "Adds a storage device to a VM.",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddStorageDeviceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Storage device added successfully"
          },
          "404": {
            "description": "VM not found"
          }
        }
      }
    },
...
    "/api/v1/jobs/storage/{jobId}": {
      "get": {
        "tags": ["Jobs"],
        "summary": "Get Storage Job",
        "description": "Gets details of a specific storage job.",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorageJobResponse"
                }
              }
            }
          },
          "404": {
            "description": "Job not found"
          }
        }
      },
      "delete": {
        "tags": ["Jobs"],
        "summary": "Delete Storage Job",
        "description": "Deletes a completed storage job.",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Storage job deleted successfully"
          },
          "404": {
            "description": "Job not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "StorageJobResponse": {
        "type": "object",
        "properties": {
          "jobId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "enum": ["Unknown", "New", "Starting", "Running", "Suspended", "ShuttingDown", "Completed", "Terminated", "Killed", "Exception", "Service"]
          },
          "percentComplete": {
            "type": "integer",
            "format": "int32"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "StorageDeviceResponse": {
        "type": "object",
        "properties": {
          "deviceId": {
            "type": "string"
          },
          "deviceType": {
            "type": "string"
          },
          "path": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "AddStorageDeviceRequest": {
        "type": "object",
        "properties": {
          "deviceType": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        },
        "required": ["deviceType", "path"]
      },
...
    }
  }
}